Merge bk://kernel.bkbits.net/gregkh/linux/driver-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/11/04 14:44:35-08:00 maneesh
[PATCH] fix kernel BUG at fs/sysfs/dir.c:20!

On Thu, Nov 04, 2004 at 12:52:38PM -0800, Greg KH wrote:
> Hi,
>
> I get the following BUG in the sysfs code when I do:
> 	- plug in a usb-serial device.
> 	- open the port with 'cat /dev/ttyUSB0'
> 	- unplug the device.
> 	- stop the 'cat' process with control-C
>
> This used to work just fine before your big sysfs changes.

There is a similar problem reported by s390 people where we see parent
kobject (directory) going away before child kobject (sub-directory). It
seems kobject code is able to handle this, but not the sysfs. What could
be happening that in sysfs_remove_dir() of parent directory, we try to
remove its contents. It works well with the regular files as it is the
final removal for sysfs_dirent corresponding to the files. But in case
of sub-directory we are doing an extra sysfs_put().  Once while removing
parent and the other one being the one from when sysfs_remove_dir() is
called for the child.

The following patch worked for the s390 people, I hope same will work in
this case also.


o Do not remove sysfs_dirents corresponding to the sub-directory in
  sysfs_remove_dir(). They will be removed in the sysfs_remove_dir() call
  for the specific sub-directory.

Signed-off-by: Maneesh Soni <maneesh@in.ibm.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 13:10:10-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleed-2.6
into kroah.com:/home/greg/linux/BK/driver-2.6

2004/11/04 12:00:36-08:00 rml
[PATCH] kobject_uevent: add MAINTAINER entry

Attached patch adds a MAINTAINER entry for the kernel event layer.


Signed-Off-By: Robert Love <rml@novell.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 12:00:12-08:00 rml
[PATCH] kobject_uevent: fix init ordering

Looks like kobject_uevent_init is executed before netlink_proto_init and
consequently always fails.  Not cool.

Attached patch switches the initialization over from core_initcall (init
level 1) to postcore_initcall (init level 2).  Netlink's initialization
is done in core_initcall, so this should fix the problem.  We should be
fine waiting until postcore_initcall.

Also a couple white space changes mixed in, because I am anal.

Signed-Off-By: Robert Love <rml@novell.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 11:59:39-08:00 greg
kevent: fix build error if CONFIG_KOBJECT_UEVENT is not selected.

Thanks to Serge Hallyn <serue@us.ibm.com> for pointing this out.

Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:57:45-08:00 tj
[PATCH] driver-model: device_add() error path reference counting fix

 df_05_device_add_ref_fix.patch

 In device_add(), @dev wan't put'd properly when it has zero length
bus_id (error path).  Fixed.


Signed-off-by: Tejun Heo <tj@home-tj.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:55:28-08:00 tj
[PATCH] driver-model: kobject_add() error path reference counting fix

 df_04_kobject_add_ref_fix.patch

In kobject_add(), @kobj wasn't put'd properly on error path.  This
patch fixes it.


Signed-off-by: Tejun Heo <tj@home-tj.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:54:31-08:00 tj
[PATCH] driver-model: sysfs_release() dangling pointer reference fix

 df_03_sysfs_release_fix.patch

Some attributes are allocated dynamically (e.g. module and device
parameters) and are usually deallocated when the assoicated kobject is
released.  So, it's not safe to access attr after putting the kobject.


Signed-off-by: Tejun Heo <tj@home-tj.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:48:20-08:00 tj
[PATCH] driver-model: bus_recan_devices() locking fix

 df_02_bus_rescan_devcies_fix.patch

 bus_rescan_devices() eventually calls device_attach() and thus
requires write locking the corresponding bus.  The original code just
called bus_for_each_dev() which only read locks the bus.  This patch
separates __bus_for_each_dev() and __bus_for_each_drv(), which don't
do locking themselves, out from the original functions and call them
with read lock in the original functions and with write lock in
bus_rescan_devices().


Signed-off-by: Tejun Heo <tj@home-tj.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:28:20-08:00 tj
[PATCH] driver-model: comment fix in bus.c

 df_01_driver_attach_comment_fix.patch

bus_match() was renamed to driver_probe_device() but the comment for
device_attach() wasn't updated.  This patch updates it.


Signed-off-by: Tejun Heo <tj@home-tj.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/04 10:12:44-08:00 kay.sievers
[PATCH] add the physical device and the bus to the hotplug environment

Add the sysfs path of the physical device to the hotplug event of class
and block devices. This should solve the userspace issue not to know if
the device is a virtual one and the "device" symlink will never be created,
but we sit there and wait for it to show up not knowing when we should
give up.

Also the bus name is added to the hotplug event, so we don't need to
reverse lookup in the /sys/bus/* directory which bus our physical
device belongs to. This is e.g. the value matched against the BUS= key,
that may be used in an udev rule.

This is a PCI network card:
  ACTION=add
  SUBSYSTEM=net
  DEVPATH=/class/net/eth0
  PHYSDEVPATH=/devices/pci0000:00/0000:00:1e.0/0000:02:01.0
  PHYSDEVBUS=pci
  INTERFACE=eth0
  SEQNUM=827
  PATH=/sbin:/bin:/usr/sbin:/usr/bin
  HOME=/

This is a IDE CDROM:
  ACTION=add
  SUBSYSTEM=block
  DEVPATH=/block/hdc
  PHYSDEVPATH=/devices/pci0000:00/0000:00:1f.1/ide1/1.0
  PHYSDEVBUS=ide
  SEQNUM=1017
  PATH=/sbin:/bin:/usr/sbin:/usr/bin
  HOME=/

This is an USB-stick partition:
  ACTION=add
  SUBSYSTEM=block
  DEVPATH=/block/sda/sda1
  PHYSDEVPATH=/devices/pci0000:00/0000:00:1d.1/usb3/3-1/3-1:1.0/host1/target1:0:0/1:0:0:0
  PHYSDEVBUS=scsi
  SEQNUM=1032
  PATH=/sbin:/bin:/usr/sbin:/usr/bin
  HOME=/


Signed-off-by: Kay Sievers <kay.sievers@vrfy.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/03 13:50:47-08:00 bunk
[PATCH] small sysfs cleanups

The patch below does the following cleanups for the sysfs code:
- remove the unused global function sysfs_mknod
- make some structs and functions static

Please check whether this patch is correct, or whether some of the
things I made static should be used globally in the forseeable future.


Signed-off-by: Adrian Bunk <bunk@stusta.de>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

2004/11/03 13:50:35-08:00 maneesh
[PATCH] sysfs: fix sysfs backing store error path confusion

o sysfs_new_dirent to retrun 0 if kmalloc fails. Thanks to Milton Miller
  for spotting this.

Signed-off-by: Maneesh Soni <maneesh@in.ibm.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

BKrev: 418ad2448OVBp8uX8WANS92SOCwaOQ