pc110 touchpad driver: be more polite
The PC110 touchpad driver used to just assume that the hardware
exists, even though in fact the hw is extremely rare indeed, and
just requesting all the resources might stomp on some _other_
hardware.

Try to minimize the damage by realizing that the touchpad hw only
exists on old ISA-only hardware, and if we have found a PCI device,
we should not try to load the touchpad driver.

Verified to fix things for Andries Brouwer.

BKrev: 418ea525VLLvKi-YiyUijHNoc0HtMQ