[PATCH] ppc32: CPM2 bug
Rune has determined that:
res6 should be 94 bytes long.

However... It does work @ 92 bytes because the compiler alligns the next
struct (pci_cpm2_t) on a int boundary anyways (no packing specified fo
the structs)

I tested it with a module that pronted out the address of the PCI struct
that follows the sit_cpm2_t struct.

Imap = 0xF0000000
With res6[92]
im-pci is @  0xf0010430

With res6[94]
im-pci is @  0xf0010430

So we're saved by the compiler.

Here's the patch

Signed-off-by: Rune Torgersen <runet@innovsys.com>
Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418eea91jrgqqfYD9szQtVRnY5VwVg