[PATCH] uml: add startup check for mmap(...PROT_EXEC...) from /tmp.
This adds a check that /tmp is not mounted noexec.  UML needs to be able to do
PROT_EXEC mmaps of temp files.  Previously, a noexec /tmp would cause an early
mysterious UML crash.  (Actually, not /tmp but $TMP, which often is the same).

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418eebc42p4uzPQwo3_m0AUl_Nq40A