[PATCH] uml: use kallsyms when dumping stack
This patch makes UML use print_symbol() when dumping the stack.  It is not yet
complete (the i386 version uses the frame pointers to walk the stack, for
instance), but should be at least enough.

Also, it uses bust_spinlock() when dumping errors to help them reach the
console.  This appears not to be enough - however I analyzed the UML console
code, and it seems that UML console don't get flushed *at all* when exiting.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418eee43vdM9dk7NS7NBcpWXgEItSQ