[PATCH] uml: clear singlestep
The patch syscall-security-3 is wrong (provided by me, sorry).  I missed,
that singlestepping_skas() used to reset PT_DTRACE.  This was handled
differently in tt and skas.  With syscall-security-3 applied, a process in
SKAS that singlestepped once continues to singlestep until the next
systemcall occurs, even if it is resumed with PTRACE_CONT or
PTRACE_SYSCALL.

This fix unifies the usage of PT_DTRACE in TT and SKAS.  PT_DTRACE now is
set by ptrace(PTRACE_SINGLESTEP,...) and reset by singlestepping() and it
is evaluated in kern_do_signal().

Signed-off-by: Bodo Stroesser <bstroesser@fujitsu-siemens.com>
Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418eeef3g9_nYSidyWB9N9MkwJuh9w