[PATCH] remove journal callback code from jbd
With the demise of intermezzo, the journal callback stuff in jbd is
entirely unused (neither ext3 nor ocfs2 use it), and thus will only bitrot
and bloat the kernel with code and datastructure growth.  If intermezzo
ever gets resurrected this will be the least of the problems they have to
face (both with generic kernel as jbd).

Signed-off-by: Arjan van de Ven <arjan@infradead.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418eefe40xB4It4Jc2NjBa7RhlI_SA