[PATCH] Fix ext3_dx_readdir
When there are more than one entry in fname linked list, the current
implementation of ext3_dx_readdir() can not traverse all entries correctly
in the case that call_filldir() fails.

If we use system call readdir() to read entries in a directory which
happens that "." and ".." in the same fname linked list.  Each time we call
readdir(), it will return the "." entry and never returns 0 which indicates
that all entries are read.

Although chances that more than one entry are in one fname linked list are
very slim, it does exist.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418ef121P6KaSimQckr18xD7P_b3bg