[NETLINK]: Fix mc_list operations
While doing the 2.4 version of the netlink patch, I noticed a bug
in my mc_list code.  First of all it wasn't holding the table lock
when adding nodes to the list.  It also didn't take the node off
the list when multicast is switched off.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 41993483G2Ore7atl9S1KToDcnwV9Q