[PATCH] uml: fix signal mask on delivery error
From: Bodo Stroesser <bstroesser@fujitsu-siemens.com>

If the user stack limit is reached or the signal stack assigned with
sigaltstack() is invalid when a user signal handler with SA_ONSTACK has to be
started, the signal mask of the interrupted user program is modified.  This
happens because the mask, that should be used with the handler only, is
written to "current->blocked" even if the handler could not be started.  But
without a handler, no rewrite of the original mask at sys_sigreturn will be
done.

A slightly different case is sys_sigsuspend(), where the mask is already
modified when kern_do_signal() is started.  "*oldset" and "current->blocked"
are not equal here and thus current->blocked has to be set to *oldset, if an
error occurs in handle_signal().

For both cases I've written small tests, and with the patch the result is OK.
This issue is relevant for other architectures too (e.g.  i386, I've seen).

Signed-off-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41997bebND-gFX2Duw8PxchVihoeqw