[PATCH] Don't remove /sys in initramfs
Using the "resume" kernel parameter together with an initramfs revealed a
bug that causes removal of the /sys directory in the initramfs' tmpfs,
making the system unbootable.

The source of the problem is that the try_name() function removes the /sys
directory unconditionally, instead of removing it only when it has been
created by try_name().

The attached patch only removes /sys if it has been created before.

Signed-off-by: Juerg Billeter <juerg@paldo.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41997cdeitrNbCR_J5anqp5s2-dpIw