[PATCH] Fix pmac_zilog.c so it compiles again
It seems that pmac_zilog.c got missed in the dev.power_state to
dev.power.power_state conversion.  This patch makes that change, and also
fixes a problem where it would not compile if CONFIG_MAGIC_SYSRQ was set but
CONFIG_SERIAL_CORE_CONSOLE was not.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41997d05fdBDWBPohci46GomzVBeCQ