[PATCH] Multilink fix for ppp_generic.c
I released ppp-2.4.3 yesterday, with much improved support for multilink
operation - when the first link is terminated, its pppd no longer exits
immediately, but keeps running in order to keep the ppp interface up while
there are other links still in the bundle.

However, this shows up a bug in the kernel ppp driver, which is that there is
no way for the pppd controlling the bundle to know when the last link in the
bundle is terminated.  This patch provides such a way: with this patch, pppd
will get an EOF when reading from the /dev/ppp instance for the bundle when
there are no channels connected.

The change does not affect older versions of pppd or normal non-multilink
operation (I have tested to make sure of that).

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41997d44xSh4W0FLG7qb04Pq3CmXGw