Merge http://linux-mh.bkbits.net/bluetooth-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/11/17 11:21:33+01:00 marcel
[Bluetooth] Check for L2CAP reliability

The L2CAP sockets can now set LM_RELIABLE flag and get notification
when a reliability problem with the ACL connection is detected. The
Bluetooth qualification tests require this functionality.

Signed-off-by: Maxim Krasnyansky <maxk@qualcomm.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/11/17 10:38:50+01:00 marcel
[Bluetooth] Add support for L2CAP secure mode

The secure mode requests an authentication and enables the encryption
for that connection. After the successful execution of these procedures
it triggers the generation of a new link key.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/11/17 10:30:12+01:00 marcel
[Bluetooth] Correct locking for zero SCID responses

The l2cap_get_chan_by_ident() function must lock the socket with the
bh_lock_sock() function on success.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/11/17 10:25:36+01:00 marcel
[Bluetooth] Update copyright information

The initial firmware loading routine for the Broadcom based chips was
written by Maxim Krasnyansky and so he also owns a copyright for this
driver.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

BKrev: 419d15b6_Yai0SvK0h3U3Bk6mESc8g